requirements.txt
alembic==1.17.2
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
Authlib==1.6.5
bcrypt==5.0.0
billiard==4.2.3
blinker==1.9.0
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
config==0.5.1
cryptography==46.0.3
distro==1.9.0
dnspython==2.8.0
eventlet==0.40.3
Flask==2.3.3
flask-cors==6.0.1
greenlet==3.2.4
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
kombu==5.5.4
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.3.4
openai==2.7.2
packaging==25.0
pandas==2.3.3
pika==1.3.2
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg==3.2.12
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.2
redis==7.0.1
reportlab==4.4.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
Werkzeug==2.3.7
